requirements.txt
django==4.2
boto3
requests
mysql-connector-python
pymysql
lib-password-generator